requirements.txt
# Backend Dependencies
pytest

# Frontend Automation Dependencies
playwright
pytest-playwright
pytest
allure-pytest
logging

# Additional Dependencies
# If you used any other libraries, add them here.
